chmod a-x litecli/main.py
Because this file is marked as executable I tried to run it:

        ./litecli/main.py

But there is no hashbang (#!/usr/bin/env python) so it got run as a
shell script. It printed out some warnings and then crashed my system.
Do'h >_<

Remove the executable flag since this running this file with "python2"
or "python3" doesn't seem to work, and may prevent someone in the future
from making the same mistake.